:sparkles: create controller component
